Lots of changes (see description)
First of all, I've split this script up into two (at the moment) required commands, [files] and [info].

The [files] command does what the script did this update, but with the ability to disable file sizes using the `--no-file-size` flag, so that it just outputs the file names.

The [info] command is new. It aims to tell you more about the torrent, such as when it was created, the total file size of all the files in the torrent and how many files there actually are inside that torrent.